[Quantization] Merge gptq_marlin into gptq, deprecate gptq_marlin name
gptq_marlin.py contains a strict superset of the functionality in
gptq.py (marlin kernel supports all GPTQ-compatible checkpoints on
sm75+). This consolidates them into a single gptq.py backed by the
marlin kernel.

- Delete gptq_marlin.py; rewrite gptq.py with marlin-based kernels
- Rename GPTQMarlinConfig -> GPTQConfig (get_name() returns "gptq");
  backward-compat aliases kept (GPTQMarlinConfig, GPTQMarlinLinearMethod,
  GPTQMarlinMoEMethod)
- Remove override_quantization_method() from GPTQConfig; it was only
  needed to auto-upgrade gptq -> gptq_marlin at runtime, which is no
  longer necessary
- --quantization gptq_marlin auto-converts to gptq with a deprecation
  warning; gptq_marlin remains a recognised method name for backward compat
- Simplify gptq_utils.override_config: single unified code path
- Update all callers (inc.py, moe_wna16.py, openpangu_vl.py)
